requirements.txt
openslide-python
